server/requirements.txt
bert-extractive-summarizer==0.10.1
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
cymem==2.0.6
en-core-web-sm==3.2.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
keybert==0.5.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rich==12.0.0
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
Werkzeug==2.0.3
